move diff utility and test import with ci-run-diff
