+++ .github/parm/use_case_groups.json
@@ -17,7 +17,7 @@
   {
     "category": "air_quality_and_comp",
     "index_list": "0",
-    "run": false
+    "run": true
   },
   {
     "category": "climate",
